feat: adicionar validações em DTOs e padronizar mensagens de erro
- Implementa validações com Bean Validation nos DTOs:
  - Campos obrigatórios
  - Formatos específicos (CPF, placa, datas)
  - Validações customizadas (@AssertTrue)
- Padroniza todas as mensagens de erro como frases completas:
  - Formato lista com bullets para múltiplos erros
  - Capitalização automática de nomes de campos
  - Mensagens humanizadas para regras de negócio
- Atualiza GlobalExceptionHandler para o novo formato

+++ caronas/src/main/java/com/beebee/caronas/entities/Avaliacao.java
@@ -21,10 +21,10 @@ public class Avaliacao {
     private LocalDate data;
 
     private String comentarioMotorista;
-    private Double notaMotorista;
+    private Integer notaMotorista;
 
     private String comentarioCaronista;
-    private Double notaCaronista;
+    private Integer notaCaronista;
 
     @ManyToOne
     @JoinColumn(name = "idViagemAluno")
